+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.1.0
 
-* Rather big update to the way using git info. Instead of storing it in and retrieving it from workspace config, it is now retrieved with git command when needed.
-* 
+* Code optimisations.
+* Updated some text on user selection buttons.
 
 ## 0.0.9